requirements.txt
flask
torch
torchvision
pillow
opencv-python
numpy
tqdm